Planner(2B DINOv3+DA3): feature-probe + split viz toolkit; DINOv3 teacher bf16
Add the qualitative-viz toolkit for the 2B DINOv3+DA3 LIBERO planner:
- train_feature_probes.py: FeatureUpsampler (FeatUp-style DINOv3 upsampler) +
  MiniDPTProbe (single-layer DA3->depth, SILog+grad, distilled from DA3-full).
- visualize_qwen3vl2b_dinov3_da3_split.py: per-camera split viz (main/wrist,
  square panels) — DINO hi-res PCA / depth probe_v2 / DA3-full GT rows.
- diag_depth_single_layer.py, diag_v2_compare.py: depth single-layer-bottleneck
  diagnostics.
- README_probes_viz.md: toolkit + probe-checkpoint locations (binaries kept on
  Ola / local outputs/ backup, not git).
Also: DINOv3 teacher default dtype -> bf16 (frozen teacher; halves ViT-H+ fwd).
Scripts target the k1_zero2 (Ola) trainer; see README.

+++ qwen3_vl_semantic_planner/dinov3_da3_2b/dinov3_target.py
@@ -43,7 +43,7 @@ class Dinov3TargetEncoder(nn.Module):
         *,
         input_size: int = 256,
         device: str | torch.device = "cuda",
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.bfloat16,  # frozen teacher; bf16 halves ViT-H+ fwd on H100
     ) -> None:
         super().__init__()
         try:
